separate error handling & merge security and websocket config
