+++ build.gradle.kts
@@ -23,8 +23,6 @@ extra["spring-security.version"]="6.0.2"
 extra["snippetsDir"] = file("build/generated-snippets")
 
 dependencies {
-	//implementation("org.springframework.security:spring-security-web:6.0.2")
-	//implementation("org.springframework.security:spring-security-config:6.0.2")
 	implementation("org.springframework:spring-messaging")
 	implementation("org.springframework.security:spring-security-messaging:6.0.2")
 	implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.0.1")
