+++ src/main/kotlin/com/wafflestudio/webgam/global/websocket/controller/WebSocketController.kt
@@ -12,7 +12,6 @@ import com.wafflestudio.webgam.global.security.jwt.JwtProvider
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.messaging.handler.annotation.*
-import org.springframework.messaging.simp.annotation.SendToUser
 import com.wafflestudio.webgam.global.security.model.UserPrincipal
 import com.wafflestudio.webgam.global.websocket.dto.WebSocketDto
 import org.springframework.stereotype.Controller
@@ -50,14 +49,6 @@ class WebSocketController (
 
         val sentMessage = WebSocketDto<String>((auth.principal as UserPrincipal).user, message.content)
         return sentMessage
-    }
-
-    @MessageExceptionHandler
-    @SendToUser("/queue/errors")
-    fun handleException(exception: Throwable, userPrincipal: UserPrincipal): WebSocketDto<String?>{
-        logger.info("Controller /queue/errors")
-        logger.info(exception.message)
-        return WebSocketDto<String?>(userPrincipal.user, exception.message)
     }
     
     @MessageMapping("/project/{projectId}/patch.project")
